bin/cost-explorer-reports/requirements.txt
-i https://pypi.org/simple
boto3==1.26.27
botocore==1.29.27 ; python_version >= '3.7'
click==8.1.3
jmespath==1.0.1 ; python_version >= '3.7'
prettytable==3.5.0
python-dateutil==2.8.2 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
s3transfer==0.6.0 ; python_version >= '3.7'
six==1.16.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
urllib3==1.26.13 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
wcwidth==0.2.5
fpdf2==2.6.1
